Consolidated spec files into one
