+++ cypress/e2e/functional/portal_user_tests/remove_teacher_comment_spec.js
@@ -15,6 +15,10 @@ const clueTeacher2 = {
   username: "clueteachertest2",
   password: "password"
 };
+const queryParams = {
+  teacher7QaNetworkQueryParams: "/?unit=https://models-resources.concord.org/clue-curriculum/branch/add-test-unit-qa/qa/content.json&problem=0.1&appMode=qa&demoName=add-test-unit-qa&fakeClass=5&fakeUser=teacher:7&network=foo",
+  teacher7MsaNetworkQueryParams: "/?appMode=qa&fakeClass=5&fakeOffering=5&problem=2.1&fakeUser=teacher:7&unit=msa&network=foo"
+};
 const ss = [{ "section": "problems",
               "subsection": "Introduction",
               "sectionCode": "introduction" },
@@ -23,7 +27,29 @@ const ss = [{ "section": "problems",
               "sectionCode": "initialChallenge" }
             ];
 
-function beforeTest(url, clueTeacher, reportUrl) {
+const ss1 = [{"section": "problems",
+              "subsection": "Introduction",
+              "sectionCode": "introduction" },
+              { "section": "problems",
+              "subsection": "Initial Challenge",
+              "sectionCode": "initialChallenge" },
+              { "section": "problems",
+              "subsection": "What If...?",
+              "sectionCode": "whatIf" },
+              { "section": "problems",
+              "subsection": "Now What Do You Know?",
+              "sectionCode": "nowWhatDoYouKnow" },
+              { "section": "teacher-guide",
+              "subsection": "Launch",
+              "sectionCode": "launch" },
+              { "section": "teacher-guide",
+              "subsection": "Explore",
+              "sectionCode": "explore" },
+              { "section": "teacher-guide",
+              "subsection": "Summarize",
+              "sectionCode": "summarize" }
+            ];
+function beforePortalTest(url, clueTeacher, reportUrl) {
   cy.login(url, clueTeacher);
   cy.launchReport(reportUrl);
   cy.waitForLoad();
@@ -32,9 +58,19 @@ function beforeTest(url, clueTeacher, reportUrl) {
   cy.wait(10000);
 }
 
+function beforeTest(params) {
+  cy.clearQAData('all');
+  cy.visit(params);
+  cy.waitForLoad();
+  cy.openTopTab("problems");
+  chatPanel.getChatPanelToggle().should('exist');
+  chatPanel.getChatPanelToggle().click();
+  cy.wait(10000);
+}
+
 describe('Delete Teacher Comments In chat panel', () => {
-  it("Remove teacher comments in chat panel", () => {
-    beforeTest(portalUrl, clueTeacher1, reportUrl1);
+  it("Delete teacher comments in the chat panel on the Learn portal", () => {
+    beforePortalTest(portalUrl, clueTeacher1, reportUrl1);
 
     ss.forEach(tab => {
       cy.openTopTab(tab.section);
@@ -50,7 +86,7 @@ describe('Delete Teacher Comments In chat panel', () => {
     
     cy.log("login teacher2 and setup clue chat");
     cy.logout(portalUrl);
-    beforeTest(portalUrl, clueTeacher2, reportUrl2);
+    beforePortalTest(portalUrl, clueTeacher2, reportUrl2);
 
     ss.forEach(tab => {
       cy.openTopTab(tab.section);
@@ -62,7 +98,36 @@ describe('Delete Teacher Comments In chat panel', () => {
       cy.clickProblemResourceTile(tab.sectionCode);
       // Teacher 2 tile comment
       chatPanel.deleteTeacherComments();
-    });
+    });  
+  });
+  it('Delete chat panel comment tags', () => {
+    beforeTest(queryParams.teacher7QaNetworkQueryParams);
+
+    cy.log('Delete comment tags on document comment');
+    cy.openTopTab("problems");
+    cy.openProblemSection("Introduction");
+    chatPanel.deleteTeacherComments();
     
+    cy.log('Delete comment tags on tile comment');
+    cy.openTopTab("problems");
+    cy.clickProblemResourceTile('introduction');
+    chatPanel.deleteTeacherComments();
+  });
+  it('Delete chat panel comments in subtabs', () => {
+    beforeTest(queryParams.teacher7MsaNetworkQueryParams);
+
+    ss1.forEach(tab => {
+      cy.openTopTab(tab.section);
+      cy.openProblemSection(tab.subsection);
+      cy.wait(2000);
+      // document comment
+      cy.log('Delete comments on document comment');
+      chatPanel.deleteTeacherComments();
+      // click first tile
+      cy.log('Delete comments on tile comment');
+      cy.clickProblemResourceTile(tab.sectionCode);
+      // tile comment
+      chatPanel.deleteTeacherComments();
+    });
   });
 });
